requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==25.1.0
mysqlclient==2.2.8
packaging==26.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
